+++ src/commands/aws/permission-set.ts
@@ -14,8 +14,8 @@ import { getAwsConfig } from "../../plugins/aws/config";
 import { assumeRoleWithIdc } from "../../plugins/aws/idc";
 import { AwsCredentials } from "../../plugins/aws/types";
 import { Authn } from "../../types/identity";
+import { provisionRequest } from "../shared/request";
 import { AssumeCommandArgs, AssumePermissionSetCommandArgs } from "./types";
-import { ACCESS_EXISTS_ERROR_MESSAGE, provisionRequest } from "./util";
 import { pick } from "lodash";
 import { sys } from "typescript";
 import yargs from "yargs";
