+++ src/commands/aws/role.ts
@@ -13,8 +13,8 @@ import { print1, print2 } from "../../drivers/stdio";
 import { AwsCredentials } from "../../plugins/aws/types";
 import { assumeRoleWithOktaSaml } from "../../plugins/okta/aws";
 import { Authn } from "../../types/identity";
+import { provisionRequest } from "../shared/request";
 import { AssumeRoleCommandArgs } from "./types";
-import { ACCESS_EXISTS_ERROR_MESSAGE, provisionRequest } from "./util";
 import { pick } from "lodash";
 import { sys } from "typescript";
 import yargs from "yargs";
